added check on update method

+++ backend/api/boatController.py
@@ -1,6 +1,6 @@
 from flask import request, jsonify
 
-from api.resource import ApiResource
+from api.resource import ApiResource, checkAdminRight
 from domain.boat import BoatRequest
 from infra.boatRepository import BoatRepository
 
@@ -12,6 +12,7 @@ class BoatController(ApiResource):
         return "/boat"
 
     def post(self):
+        checkAdminRight()
         boat_json = request.get_json()
         boat_request = BoatRequest(boat_json)
         boat_repository.create(boat_request)
@@ -31,6 +32,8 @@ class BoatByIdController(ApiResource):
         return jsonify(boat.__dict__())
 
     def put(self, id):
+        checkAdminRight()
+
         boat_json = request.get_json()
 
         boat_request = BoatRequest(boat_json)
@@ -39,5 +42,7 @@ class BoatByIdController(ApiResource):
         return jsonify({"success": True})
 
     def delete(self, id):
+        checkAdminRight()
+
         boat_repository.delete(id)
         return jsonify({"success": True})
